requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
streamlit
joblib
python-pptx
scipy
